Change architecture for buffering actions
Previously, actions were buffered on a per-actor basis. This meant, for example,
that updating the usernote wiki page with notes from n different rules meant
hitting the wiki API O(n) times. With this change, actions of a given kind are
buffered on a per-subreddit basis, and can be handled all at once. The number of
API requests in the example would therefore be constant in the number of rules.