Allow multiple moderators to target the same item
Previously, only one action could target a given item in order to prevent
repetition when items re-enter the modqueue. This commit relaxes the rule by
allowing multiple actions provided they come from different moderators.

This has the side-effect of possibly creating redundant or conflicting actions
on the same post; for example, two different moderators posting notifications.
However, as actually deployed in the philosophy subreddits there don't seem to
be enough overlapping moderator actions that this is likely to be a frequent
problem.